Move keying material exporter check from syshead.h to configure.ac
It turns out that openvpnserv also includes syshead.h, and has
ENABLE_CRYPTO_* defined in it's config.h, but doesn't have the necessary
CFLAGS / LIBS to actualy compile and link against the crypto libraries.
That of course breaks Windows builds.

To fix this, change the compile-time check in syshead.h into a
configure-time check in configure.ac. That's more consistent with how we
do other feature checks anyway.

Signed-off-by: Steffan Karger <steffan.karger@foxcrypto.com>